backend/requirements.txt
# Decision Platform Backend Dependencies

# FastAPI and API framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.2.0  # Required for Pydantic EmailStr

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0  # PostgreSQL async driver
psycopg2-binary==2.9.9  # PostgreSQL sync driver

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Caching and Task Queue
redis==5.0.1
celery==5.3.4

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
tensorflow==2.15.0
torch==2.1.2
xgboost==2.0.2

# Model Explainability
shap==0.44.0
lime==0.2.0.1

# ML Experiment Tracking
mlflow==2.9.2

# Data Processing
openpyxl==3.1.2  # Excel support
pyarrow==14.0.2  # Parquet support
requests==2.31.0  # HTTP requests

# Cloud Providers
boto3==1.34.0  # AWS
google-cloud-storage==2.10.0  # GCP
azure-storage-blob==12.19.0  # Azure

# Monitoring and Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.40.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI
factory-boy==3.3.0  # Test data factories

# Code Quality
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.8.0

# Environment and Configuration
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
